Reconcile Model-B code clobbered by merge c1c6e2d; widen university_id to BIGINT
Merge c1c6e2d resolved its conflicts in favour of an older pre-migration
branch, silently reverting the Model-B migration. That left 16 backend
files uncompilable (174 errors) and a frontend that failed to compile
because services/api.js lost 13 exports.

Restores the Model-B implementations from 29d7b53 and keeps only the
genuinely newer additions the old branch happened to match: actuator URL
rules, @EnableCaching, country caching, the findAll(Pageable) overloads,
and the StudentController COMPANY scoping plus user auto-provisioning.
Re-adds the /api/students portal endpoints and the day-diary comment
fields the UI relies on, and drops the duplicated legacy seeder.

Also standardises universityId on java.lang.Long across University,
School, Department and Programme so every university_id column is BIGINT.
The mysql profile runs ddl-auto=none, so the widening ships as the
idempotent backend/migration/widen_university_id_bigint.sql and is
mirrored in schema.sql and catalog_seed.sql. Widening INT to BIGINT is
non-destructive and no FK is declared on any university_id column, so it
is safe to apply to production after a backup.

Backend: 41/41 tests pass. Frontend: build succeeds, 7/7 tests pass.

+++ backend/migration/catalog_seed.sql
@@ -8,7 +8,7 @@ START TRANSACTION;
 CREATE TABLE IF NOT EXISTS schools (
     parent_school_id integer,
     school_id integer not null,
-    university_id integer not null,
+    university_id bigint not null,
     school_code varchar(255),
     school_name varchar(255) not null,
     type varchar(255),
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS schools (
 CREATE TABLE IF NOT EXISTS departments (
     department_id integer not null,
     school_id integer not null,
-    university_id integer not null,
+    university_id bigint not null,
     department_name varchar(255) not null,
     primary key (department_id)
 ) engine=InnoDB;
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS programmes (
     duration_years integer not null,
     programme_id integer not null,
     school_id integer not null,
-    university_id integer not null,
+    university_id bigint not null,
     programme_code varchar(255) not null,
     programme_level varchar(255) not null,
     programme_name varchar(255) not null,
