+++ backend/src/main/java/com/example/demo/department/DepartmentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface DepartmentRepository extends JpaRepository<Department, Integer> {
     List<Department> findBySchoolId(Integer schoolId);
-    List<Department> findByUniversityId(Integer universityId);
+    List<Department> findByUniversityId(Long universityId);
 }
