+++ backend/src/main/java/com/example/demo/programme/Programme.java
@@ -17,7 +17,7 @@ public class Programme {
     private Integer schoolId;
 
     @Column(name = "university_id", nullable = false)
-    private Integer universityId;
+    private Long universityId;
 
     @Column(name = "department_id", nullable = true)
     private Integer departmentId;
@@ -53,11 +53,11 @@ public class Programme {
         this.schoolId = schoolId;
     }
 
-    public Integer getUniversityId() {
+    public Long getUniversityId() {
         return universityId;
     }
 
-    public void setUniversityId(Integer universityId) {
+    public void setUniversityId(Long universityId) {
         this.universityId = universityId;
     }
